WIP
This is a sketch I did for how we might approach the problem of pre
fetching websites to speed up the application of RawInput records.

This code is not tested and has never been run - it was mostly a thought
experiment around ways to make the network requests in parallel while
the CSV is being parsed.